+++ integration_tests/tests/flowmachine_server_tests/test_helper_functions.py
@@ -34,6 +34,6 @@ def test_send_zmq_message_and_receive_reply(zmq_host, zmq_port):
     reply = send_zmq_message_and_receive_reply(
         FM_EXAMPLE_MESSAGE, host=zmq_host, port=zmq_port
     )
-    assert "e39b0d45bc6b46b7700c67cd52f00455" == reply["payload"]["query_id"]
+    assert "a2fb1efca05a42d558e8c613970262de" == reply["payload"]["query_id"]
     # assert reply["status"] in ("executing", "queued", "completed")
     assert reply["status"] in ("success")